fix: Issue with polymorphic models fix and question behaviour.
The previous fix for the issue with polymorphic models introduced problems for BaseQuestion models with mixins (Mixins were ignored). As a consequence, question items and question scales were not generated properly and hence the questionnaire did not render as expected. This was fixed by overwriting the delete method on the DonationProject model to delete related polymorphic models manually before calling the standard collector route.